Add files via upload
Added TwilightTimes.xlsx which contains a table of twilight times for sunrise and sunset, and 6, 12, and 18 deg twilight in the evening and morning.  Includes data for every day between Jan 1 2000 and Dec 31 2020.  Times are stored as fractions of a day, and all are in UTC.  Data were taken from the public USNO website, at http://aa.usno.navy.mil/data/docs/RS_OneYear.php